rename org.jboss.da.rest.spi to api
