+++ reports-rest/src/main/java/org/jboss/da/rest/listings/BlackListImpl.java
@@ -17,7 +17,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-import org.jboss.da.rest.spi.BlackList;
+import org.jboss.da.rest.api.BlackList;
 
 /**
  *
